+++ Source/WebKit/qt/WebCoreSupport/QWebPageAdapter.h
@@ -352,7 +352,7 @@ public:
     };
 
     ViewportAttributes viewportAttributesForSize(const QSize& availableSize, const QSize& deviceSize) const;
-    void setDevicePixelRatio(int devicePixelRatio);
+    void setDevicePixelRatio(float devicePixelRatio);
 
     QWebSettings *settings;
